[Converter] Rename const into pure and drop constF and idF
